% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr.R
\name{cr_journals}
\alias{cr_journals}
\title{Search Crossref journals}
\usage{
cr_journals(query, rows = 10)
}
\arguments{
\item{query}{Search query string (journal title)}

\item{rows}{Number of results. Default 10.}
}
\value{
tibble: issn, title, publisher, subjects, total_dois
}
\description{
Search Crossref journals
}
